+++ apm-sniffer/apm-sdk-plugin/spring-plugins/mvc-annotation-5.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/spring/mvc/v5/define/AbstractSpring5Instrumentation.java
@@ -20,7 +20,7 @@ package org.apache.skywalking.apm.plugin.spring.mvc.v5.define;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.ClassInstanceMethodsEnhancePluginDefine;
 
 public abstract class AbstractSpring5Instrumentation extends ClassInstanceMethodsEnhancePluginDefine {
-    public static final String WITNESS_CLASSES = "org.springframework.beans.annotation.AnnotationBeanUtils";
+    public static final String WITNESS_CLASSES = "org.springframework.core.ReactiveAdapterRegistry";
 
     @Override
     protected final String[] witnessClasses() {