+++ CHANGES.md
@@ -138,6 +138,11 @@ Callable {
 * Fix the thread safety bug of finishing operation for the span named "SpringCloudGateway/sendRequest"
 * Fix NPE in guava-eventbus-plugin.
 * Add WebSphere Liberty 23.x plugin
+* Add Plugin to support aerospike Java client
+* Add ClickHouse parsing to the jdbc-common plugin.
+* Support to trace redisson lock
+* Upgrade netty-codec-http2 to 4.1.94.Final
+* Upgrade guava to 32.0.1
 
 #### Documentation